% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Input.input}
\alias{Input.input}
\title{Input.input Object}
\usage{
Input.input(csvInstance = NULL)
}
\arguments{
\item{csvInstance}{No description}
}
\value{
Input.input object
}
\description{
Input.input Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Input functions: \code{\link{Input}},
  \code{\link{hostedmodels.predict}}
}
