% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Input}
\alias{Input}
\title{Prediction API Objects 
Lets you access a cloud hosted machine learning service that makes it easy to build smart apps}
\usage{
Input(Input.input = NULL, input = NULL)
}
\arguments{
\item{Input.input}{The \link{Input.input} object or list of objects}

\item{input}{No description}
}
\value{
Input object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:05:37
filename: /Users/mark/dev/R/autoGoogleAPI/googlepredictionv12.auto/R/prediction_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Input Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Input functions: \code{\link{Input.input}},
  \code{\link{hostedmodels.predict}}
}
