% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{training.delete}
\alias{training.delete}
\title{Delete a trained model}
\usage{
training.delete(data)
}
\arguments{
\item{data}{mybucket/mydata resource in Google Storage}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/prediction
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/prediction)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/prediction/docs/developer-guide}{Google Documentation}
}

