% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training.modelInfo}
\alias{Training.modelInfo}
\title{Training.modelInfo Object}
\usage{
Training.modelInfo(classificationAccuracy = NULL, meanSquaredError = NULL,
  modelType = NULL)
}
\arguments{
\item{classificationAccuracy}{No description}

\item{meanSquaredError}{No description}

\item{modelType}{No description}
}
\value{
Training.modelInfo object
}
\description{
Training.modelInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Training functions: \code{\link{Training}},
  \code{\link{training.insert}}
}
