% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training}
\alias{Training}
\title{Training Object}
\usage{
Training(Training.modelInfo = NULL, id = NULL, modelInfo = NULL,
  selfLink = NULL, trainingStatus = NULL)
}
\arguments{
\item{Training.modelInfo}{The \link{Training.modelInfo} object or list of objects}

\item{id}{No description}

\item{modelInfo}{No description}

\item{selfLink}{No description}

\item{trainingStatus}{No description}
}
\value{
Training object
}
\description{
Training Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo}},
  \code{\link{training.insert}}
}
